+++ src/components/LoadingScreen/index.tsx
@@ -1,0 +1,9 @@
+import { Container, Text } from "./styles";
+
+export default function LoadingScreen() {
+    return (
+        <Container>
+            <Text>Loading...</Text>
+        </Container>
+    );
+}