+++ src/components/MIDIKeyboard/index.tsx
@@ -1,13 +1,18 @@
-import { BlackKey, BlackKeyBlock, BlocksArea, Container, Keyboard, WhiteKey, WhiteKeyBlock } from "./styles";
+import Soundfont from 'soundfont-player';
+import noteNames from "../../core/utils/music-utils";
+import { BlackKey, BlackKeyBlock, BlocksArea, Container, Keyboard, NoteText, WhiteKey, WhiteKeyBlock } from "./styles";
 import { INote } from "../../core/interfaces/models/note.interface";
 import { KeyboardSize } from "../../core/types/keyboard-size.type";
+import { playNote } from "../../core/utils/web-audio-api";
 import { useEffect } from "react";
 
 interface MIDIKeyboardProps {
     blackKeys: number[];
     currentTime: number;
+    instruments: {[key: number]: Soundfont.Player};
     notes: INote[];
     onMidiEnd: () => void;
+    showNotes: boolean;
     size: KeyboardSize;
     totalTime: number;
     whiteKeys: number[];
@@ -16,15 +21,23 @@ interface MIDIKeyboardProps {
 export default function MIDIKeyboard({ 
     blackKeys,
     currentTime,
+    instruments,
     notes,
     onMidiEnd,
+    showNotes,
     size,
     totalTime,
     whiteKeys
 }: MIDIKeyboardProps) {
     useEffect(() => {
-        if(currentTime > totalTime + 1000) onMidiEnd();
-    }, [currentTime, notes, onMidiEnd]);
+        if(currentTime > totalTime + 1000) {
+            onMidiEnd();
+        }
+
+        notes.filter(b => b.startsAt < currentTime && b.startsAt + 37 > currentTime && currentTime < b.endsAt).forEach(b => {
+            playNote(instruments[b.instrument], b);
+        });
+    }, [currentTime, instruments, totalTime, notes, onMidiEnd]);
 
     return (
         <Container>
@@ -32,25 +45,33 @@ export default function MIDIKeyboard({
                 {notes.map((note, key) => whiteKeys.includes(note.key) ? (
                     note.startsAt - currentTime > 0 - (note.endsAt - note.startsAt) 
                     && note.startsAt < currentTime + 4000 
-                    && <WhiteKeyBlock
-                        key={key}
-                        $color={note.color}
-                        $duration={note.endsAt - note.startsAt} 
-                        $keyboardsize={size}
-                        $position={note.startsAt - currentTime}
-                        $notenumber={note.key}
-                    />
+                    && (
+                        <WhiteKeyBlock
+                            key={key}
+                            $color={note.color}
+                            $duration={note.endsAt - note.startsAt} 
+                            $keyboardsize={size}
+                            $position={note.startsAt - currentTime}
+                            $notenumber={note.key}
+                        >
+                            {showNotes && <NoteText>{noteNames[note.key % 12]}</NoteText>}
+                        </WhiteKeyBlock>
+                    )
                 ) : (
                     note.startsAt - currentTime > 0 - (note.endsAt - note.startsAt) 
                     && note.startsAt < currentTime + 4000 
-                    && <BlackKeyBlock
-                        key={key}
-                        $color={note.color}
-                        $duration={note.endsAt - note.startsAt} 
-                        $keyboardsize={size}
-                        $position={note.startsAt - currentTime}
-                        $notenumber={note.key}
-                    />
+                    && (
+                        <BlackKeyBlock
+                            key={key}
+                            $color={note.color}
+                            $duration={note.endsAt - note.startsAt} 
+                            $keyboardsize={size}
+                            $position={note.startsAt - currentTime}
+                            $notenumber={note.key}
+                        >
+                            {showNotes && <NoteText>{noteNames[note.key % 12]}</NoteText>}
+                        </BlackKeyBlock>
+                    )
                 ))}
             </BlocksArea>
             <Keyboard>
@@ -60,7 +81,9 @@ export default function MIDIKeyboard({
                         $keyboardsize={size}
                         $notenumber={noteNumber}
                         $pressednote={notes.find(b => b.key === noteNumber && b.startsAt < currentTime && currentTime < b.endsAt)}
-                    />
+                    >
+                        {showNotes && <NoteText>{noteNames[noteNumber % 12]}</NoteText>}
+                    </WhiteKey>
                 ))}
                 {blackKeys.map((noteNumber, key) => (
                     <BlackKey 
@@ -68,7 +91,9 @@ export default function MIDIKeyboard({
                         $keyboardsize={size}
                         $notenumber={noteNumber}
                         $pressednote={notes.find(b => b.key === noteNumber && b.startsAt < currentTime && currentTime < b.endsAt)}
-                    />
+                    >
+                        {showNotes && <NoteText>{noteNames[noteNumber % 12]}</NoteText>}
+                    </BlackKey>
                 ))}
             </Keyboard>
         </Container>