+++ src/components/MIDIKeyboard/styles.ts
@@ -2,6 +2,23 @@ import styled from "styled-components";
 import { KeyboardSize } from "../../core/types/keyboard-size.type";
 import { INote } from "../../core/interfaces/models/note.interface";
 
+const size = 1.1;
+
+const whiteKeyWidth = 23 * size;
+const blackKeyWidth = 14 * size;
+const whiteKeyHeight = 150 * size;
+const blackKeyHeight = 90 * size;
+
+const whiteBlockKeyWidth = 23 * size;
+const blackBlockKeyWidth = 14 * size;
+
+const sizeOffsets = { 
+    piano: whiteKeyWidth * -12, 
+    large: whiteKeyWidth * -16, 
+    medium: whiteKeyWidth * -21, 
+    small: whiteKeyWidth * -28
+};
+
 export const Container = styled.div`
     grid-area: player;
     position: relative;
@@ -29,18 +46,19 @@ interface KeyProps {
 }
 
 function whiteKeyOffset(keyboardSize: KeyboardSize, noteNumber: number) {
-    const sizeOffsets = { piano: -276, large: -368, medium: -483, small: -644 };
     const octNote = noteNumber % 12;
     const sizeOffset = sizeOffsets[keyboardSize] || 0;
 
     const extraWhiteKeys = [2, 4, 5, 7, 9, 11];
-    const extraOffset = extraWhiteKeys.filter(i => octNote >= i).length * 23;
+    const extraOffset = extraWhiteKeys.filter(i => octNote >= i).length * whiteKeyWidth;
 
-    return Math.floor(noteNumber / 12) * 23 * 7 + sizeOffset + extraOffset;
+    return Math.floor(noteNumber / 12) * whiteKeyWidth * 7 + sizeOffset + extraOffset;
 }
 
 export const WhiteKey = styled.div.attrs<KeyProps>(props => ({
     style: {
+        width: whiteKeyWidth,
+        height: whiteKeyHeight,
         left: `${whiteKeyOffset(props.$keyboardsize, props.$notenumber)}px`,
         backgroundColor: props.$pressednote ? props.$pressednote.color : '#FFFFFF',
         boxShadow: props.$pressednote 
@@ -50,35 +68,42 @@ export const WhiteKey = styled.div.attrs<KeyProps>(props => ({
 }))`
     position: absolute;
     bottom: 0;
-    width: 23px;
-    height: 150px;
-    border: 1px solid gray;
+    border: 1px solid #444;
+    border-bottom-left-radius: 3px;
+    border-bottom-right-radius: 3px;
     z-index: 1;
+    
+    span {
+        font-size: ${whiteKeyWidth * 0.8}px;
+        color: #000;
+    }
 `;
 
 function blackKeyOffset(keyboardSize: KeyboardSize, noteNumber: number) {
-    const offsets = { piano: -276, large: -368, medium: -483, small: -644 };
     const octNote = noteNumber % 12;
-    const sizeOffset = offsets[keyboardSize] || 0;
+    const sizeOffset = sizeOffsets[keyboardSize] || 0;
 
     const conditions = [
-        { condition: octNote >= 3, value: 25 },
-        { condition: octNote >= 6, value: 43 },
-        { condition: octNote >= 8, value: 25 },
-        { condition: octNote >= 10, value: 25 }
+        { condition: octNote >= 3, value: whiteKeyWidth * 1.1 },
+        { condition: octNote >= 6, value: whiteKeyWidth * 1.83 },
+        { condition: octNote >= 8, value: whiteKeyWidth * 1.1 },
+        { condition: octNote >= 10, value: whiteKeyWidth * 1.1 }
     ];
 
     const extraOffset = conditions.reduce(
         (acc, curr) => acc + (curr.condition ? curr.value : 0),
         0
-    );
+    ) + (whiteKeyWidth * 0.65);
 
-    return Math.floor(noteNumber / 12) * 23 * 7 + sizeOffset + extraOffset;
+    return Math.floor(noteNumber / 12) * whiteKeyWidth * 7 + sizeOffset + extraOffset;
 }
 
 export const BlackKey = styled.div.attrs<KeyProps>(props => ({
     style: {
-        left: `${blackKeyOffset(props.$keyboardsize, props.$notenumber) + 15}px`,
+        width: blackKeyWidth,
+        height: blackKeyHeight,
+        left: `${blackKeyOffset(props.$keyboardsize, props.$notenumber)}px`,
+        bottom: whiteKeyHeight - blackKeyHeight,
         backgroundColor: props.$pressednote ? props.$pressednote.color : '#000000',
         boxShadow: props.$pressednote 
             ? `-5px -15px 10px white, 5px -15px 10px white, 0 -30px 30px ${props.$pressednote.color}` 
@@ -86,11 +111,14 @@ export const BlackKey = styled.div.attrs<KeyProps>(props => ({
     }
 }))`
     position: absolute;
-    bottom: 60px;
-    width: 14px;
-    height: 90px;
-    border: 1px solid gray;
+    border: 1px solid #444;
+    border-bottom-left-radius: 3px;
+    border-bottom-right-radius: 3px;
     z-index: 2;
+
+    span {
+        font-size: ${blackKeyWidth * 0.8}px;
+    }
 `;
 
 interface BlockProps {
@@ -103,7 +131,7 @@ interface BlockProps {
 
 export const Block = styled.div.attrs<BlockProps>(props => ({
     style: {
-        bottom: `${150 + (props.$position / 5)}px`,
+        bottom: `${whiteKeyHeight + (props.$position / 5)}px`,
         height: `${props.$duration / 5}px`,
         backgroundColor: props.$color
     }
@@ -115,16 +143,34 @@ export const Block = styled.div.attrs<BlockProps>(props => ({
 
 export const WhiteKeyBlock = styled(Block).attrs<BlockProps>(props => ({
     style: {
-        left: `${whiteKeyOffset(props.$keyboardsize, props.$notenumber)}px`
+        width: whiteBlockKeyWidth,
+        left: `${whiteKeyOffset(props.$keyboardsize, props.$notenumber)}px`,
+        minHeight: whiteBlockKeyWidth * 0.8
     }
 }))`
-    width: 23px;
+    span {
+        font-size: ${whiteBlockKeyWidth * 0.8}px;
+    }
 `;
 
 export const BlackKeyBlock = styled(Block).attrs<BlockProps>(props => ({
     style: {
-        left: `${blackKeyOffset(props.$keyboardsize, props.$notenumber) + 15}px`
+        width: blackBlockKeyWidth,
+        left: `${blackKeyOffset(props.$keyboardsize, props.$notenumber)}px`,
+        minHeight: blackBlockKeyWidth * 0.8
     }
 }))`
-    width: 14px;
+    span {
+        font-size: ${blackBlockKeyWidth * 0.8}px;
+    }
+`;
+
+export const NoteText = styled.span`
+    margin: 0;
+    position: absolute;
+    bottom: 0;
+    left: 50%;
+    transform: translate(-50%, 0);
+    text-align: center;
+    font-weight: 600;
 `;