+++ src/core/interfaces/models/note.interface.ts
@@ -1,5 +1,6 @@
 export interface INote {
     key: number
+    instrument: number;
     startsAt: number;
     endsAt: number;
     velocity: number;