+++ src/core/utils/music-utils.ts
@@ -1,0 +1,5 @@
+const noteNames = [
+    'C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B'
+];
+
+export default noteNames;